Corrected files for manufacture
"kikit fab pcbway" currently does not align layers and drills
for designs with a specified auxiliary axis origin.
For now, I have corrected this by making local changes to KiKit.
Files for inner layers are also named to be consistent with
outer layers, and are included in the .gbrjob file.

Thanks to @DX-MON for the invaluable tip on re-checking the
Gerber files after switching to using KiKit for file export.