MOS6581 / MOS8580 FPGA design
This is a work in progress. The gateware is designed by someone who
has never attempted a more advanced FPGA design than a "blinky", and
nothing is tested, or even simulated. It goes without saying that the
probability that anything will currently work is approximately 0.

The design aims to be 100% cycle accurate, based on information from
https://github.com/libsidplayfp/SID_schematics/wiki and
https://github.com/daglem/reSID. The SID chip is a latched design,
internally requiring a two phase non-overlapping clock. In order to
emulate this in gateware using D flip-flops, four dummy clock phases
are used - PHI2, PHI2_PHI1, PHI1, and PHI1_PHI2.

The design includes two SID cores, where the second SID core is
addressable using combinations of the pins /IO1, A5, and A8.

The analog accuracy will initially be at the level of reSID 0.16,
with the addition of parameterizable filter curves.

To make a design for a reasonably accurate SID emulation fit in an
iCE40UP5K, a few novelties have been invented:

* Combined sawtooth/triangle waveforms without lookup tables.
* Combined pulse/sawtooth/triangle waveforms without lookup tables.
* MOS6581 waveform, envelope, and filter cutoff DAC emulation without
  lookup tables.
* Parameterizable filter cutoff curves requiring only a single 16kbit
  lookup table.

TODO:

* Real-time configuration by writing magic bytes to SID registers.
* Filtering / resampling of the 1MHz audio output.
* Possibly time division multiplexing of the two SID cores in order
  to make muacm fit.
* Simulation and testing, debugging and fixing.
* Attempt to improve analog accuracy, similar to reSID 1.0.