Use RGB 2mA current drive for POT discharge
This was planned when the board was designed, however when implementing
the gateware the iCE40 techlib doc seemed to indicate that one couldn't
read from a pin which uses the SB_RGBA_DRV primitive to drive current.
So SB_IO was used for POT discharge and the RGB current drive was
forgotten about, without considering that shorting a POT pin by applying
~5V while outputting 0 causes ~50mA to be drawn into the FPGA using
normal IO.

Luckily it turns out that SB_RGBA_DRV can be combined with SB_IO. When
the current drive is turned off using RGBLEDEN or RGB0PWM / RGB1PWM, the
pin input can be read using SB_IO.

This fixes issue #9, thanks to @sorgelig for questioning the safety of
the POT discharge.